Edits (+ code edit)
I did mostly just grammar edits, but I did edit the code parts a bit. 
As such the github gist /no longer matches/ the markdown. So it's fine if you don't merge those edits, or you could edit the gist? (I have never done a gist before so idk how it works lol)
The changes I made to the code was just renaming withDrawMoney to withdrawMoney (I also changed the heading for that function) and also it didn't need to be payable.

Any feedback is welcome, hope you're doing well.